Move UI components to another folder
